requirements.txt
Django==3.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==4.6.0
psycopg2-binary==2.8.6
gunicorn==20.0.4
django-cors-headers==3.5.0
dj-database-url==0.5.0